--- packages/backend/libraries/backend-common/CHANGELOG.md
@@ -1,37 +1,0 @@
-# Changelog
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
-<!--
-## [UNRELEASED]
-
-### Added
-### Changed
-### Deprecated
-### Removed
-### Fixed
-### Security
-### Docs
--->
-
-
-
-
-## [UNRELEASED]
-
-### Added
-- Added `AppError`.
-- Added `AppErrorKind`.
-- Added `Builder`.
-- Added `BuilderAsync`.
-- Added `Controller`.
-- Added `Converter`.
-- Added `ConverterAsync`.
-- Added `Entity`.
-- Added `Spec`.
-- Added `Writable`.
-
-
-
