--- packages/backend/libraries/backend-common/src/common/domain/modules/Converter.ts
@@ -1,7 +1,0 @@
-export interface Converter<
-  TInput = unknown,
-  TOutput = unknown,
-  TContext = void,
-> {
-  convert(input: TInput, context: TContext): TOutput;
-}
