--- packages/backend/libraries/backend-common/src/common/domain/modules/ConverterAsync.ts
@@ -1,7 +1,0 @@
-import { Converter } from './Converter';
-
-export type ConverterAsync<
-  TInput = unknown,
-  TOutput = unknown,
-  TContext = void,
-> = Converter<TInput, Promise<TOutput>, TContext>;
