+++ packages/backend/libraries/backend-common/src/index.ts
@@ -4,8 +4,6 @@ import { Publisher } from './common/application/modules/Publisher';
 import { Entity } from './common/domain/models/Entity';
 import { Builder } from './common/domain/modules/Builder';
 import { BuilderAsync } from './common/domain/modules/BuilderAsync';
-import { Converter } from './common/domain/modules/Converter';
-import { ConverterAsync } from './common/domain/modules/ConverterAsync';
 import { ReportBasedSpec } from './common/domain/modules/ReportBasedSpec';
 import { Spec } from './common/domain/modules/Spec';
 import {
@@ -25,8 +23,6 @@ export type {
   BuilderAsync,
   Either,
   Handler,
-  Converter,
-  ConverterAsync,
   Entity,
   Left,
   Publisher,
